wrap_in_gsf: Convert rate to Fraction then invert
Converts the rate argument to a Fraction before inverting it to get the
Grain duration to avoid a bug where the duration comes out as zero.

The cause of that bug is that `Fraction(1/25)` is
`Fraction(5764607523034235, 144115188075855872)` because of floating
point badness, and the bottom 32 bits of that denominator (2^57) are all
zeroes, so a zero gets stored in the file, so when parsed back out by
`read_rational()` a `Fraction(0)` is returned.

sem-ver: bugfix