Refactor GSF reading classes to share header parsing
Add support for a (buffer) bytes file_data to the existing
SyncGSFBlock. The file_data_start arg is used to indicate the start
position of the buffered data in the base file.

Add a read_remaining_block method to both SyncGSFBlock and
AsyncGSFBlock that returns the buffered remainder of the block as a
SyncGSFBlock. This should only be used in blocks that can and should be
read into memory.

Factor out the header parsing methods from GSFSyncDecoderSession and
GSFAsyncDecoderSession to a new BaseGSFDecoderSession base class.
This class provides the header parsing methods that uses a SyncGSFBlock
that contains the buffered remainder of the header block from the
GSFSyncDecoderSession or GSFAsyncDecoderSession.

Unused async read methods have been removed from AsyncGSFBlock. The
remaining ones are used to read the file header, grain data etc. where
buffering is unneccessary in terms of code duplication (minor) or
requires to much memory.