Support null is_key_frame and temporal_offset
GSF version changed to 9.

This change supports cases where is_key_frame and temporal_offset are
unknown, e.g. the processor that encapsulated the media in grains was
not capable emough to extract the info. Readers are then able to
decide to process the media differently if the values are known to be
unknown.

sem-ver: api-break